# MobyPark/OpenAPI/openapi.yaml
openapi: 3.0.4
info: {}
paths:
...
  /api/LicensePlate:
    post:
      tags:
      - LicensePlate
      summary: Registers a new license plate for the current user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLicensePlateDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateLicensePlateDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateLicensePlateDto'
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '201':
          description: License plate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadLicensePlateDto'
        '409':
          description: License plate already exists
    get:
      tags:
      - LicensePlate
      summary: Retrieves all license plates for a user.
      parameters:
      - name: username
        in: query
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadLicensePlateDto'
        '403':
          description: Not authorized to view other users' plates
        '404':
          description: User not found
  /api/LicensePlate/{licensePlate}/entry:
    get:
      tags:
      - LicensePlate
      summary: Checks if a license plate is allowed to enter a parking lot.
      parameters:
      - name: licensePlate
        in: path
        required: true
        schema:
          type: string
      - name: parkingLotId
        in: query
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Entry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadLicensePlateDto'
        '404':
          description: License plate or Parking lot not found
        '409':
          description: No spots available
  /api/LicensePlate/{licensePlate}:
    put:
      tags:
      - LicensePlate
      summary: Updates an existing license plate number.
      parameters:
      - name: licensePlate
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLicensePlateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateLicensePlateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateLicensePlateRequest'
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadLicensePlateDto'
        '404':
          description: Old license plate not found
        '409':
          description: New license plate already exists
    delete:
      tags:
      - LicensePlate
      summary: Deletes a license plate.
      parameters:
      - name: licensePlate
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: License plate not found
    get:
      tags:
      - LicensePlate
      summary: Retrieves details of a specific license plate.
      parameters:
      - name: licensePlate
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Found plate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadLicensePlateDto'
        '404':
          description: Not found
...
  /api/Payments/plate/{licensePlate}:
    get:
      tags:
      - Payments
      summary: Retrieves all payments associated with a license plate.
      parameters:
      - name: licensePlate
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentModel'
        '404':
          description: No payments found
...
components:
  schemas:
...
    CreateLicensePlateDto:
      required:
      - licensePlate
      type: object
      properties:
        licensePlate:
          minLength: 1
          type: string
          description: The license plate number (e.g., AA-12-BB)
      additionalProperties: false
      description: Data required to register a new license plate.
...
    LicensePlateModel:
      required:
      - licensePlateNumber
      type: object
      properties:
        licensePlateNumber:
          minLength: 1
          type: string
      additionalProperties: false
...
    PaymentModel:
      required:
      - amount
      - createdAt
      - licensePlateNumber
      - transactionId
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        licensePlateNumber:
          minLength: 1
          type: string
        licensePlate:
          $ref: '#/components/schemas/LicensePlateModel'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        transactionId:
          type: string
          format: uuid
        transaction:
          $ref: '#/components/schemas/TransactionModel'
      additionalProperties: false
...
    ReadLicensePlateDto:
      type: object
      properties:
        licensePlate:
          type: string
          description: The license plate number
          nullable: true
        userId:
          type: integer
          description: The ID of the user who owns this plate
          format: int64
        status:
          type: string
          description: Optional status report relating to the method this plate was
            retrieved
          nullable: true
      additionalProperties: false
      description: Details of a registered license plate.
...
    TransactionModel:
      required:
      - amount
      - bank
      - issuer
      - method
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        method:
          minLength: 1
          type: string
        issuer:
          minLength: 1
          type: string
        bank:
          minLength: 1
          type: string
      additionalProperties: false
    UpdateLicensePlateRequest:
      required:
      - newLicensePlate
      - oldLicensePlate
      type: object
      properties:
        oldLicensePlate:
          minLength: 1
          type: string
          description: The current license plate number
        newLicensePlate:
          minLength: 1
          type: string
          description: The new license plate number
      additionalProperties: false
      description: Data to update an existing license plate.
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Please enter a valid token (e.g., ''Bearer eyJ...'').

        You can obtain a token by logging in or registering a new account via the
        endpoints.'
      scheme: Bearer
      bearerFormat: JWT
security:
- Bearer: []